Extracting generation of SendTransaction examples in tests
