+++ packages/sanes-chrome-extension/src/routes/tx-request/index.stories.tsx
@@ -1,14 +1,3 @@
-import {
-  Address,
-  Algorithm,
-  Amount,
-  ChainId,
-  PublicIdentity,
-  PublicKeyBytes,
-  SendTransaction,
-  TokenTicker,
-} from '@iov/bcp';
-import { Encoding } from '@iov/encoding';
 import { action } from '@storybook/addon-actions';
 import { linkTo } from '@storybook/addon-links';
 import { storiesOf } from '@storybook/react';
@@ -22,39 +11,11 @@ import { CHROME_EXTENSION_ROOT } from '../../utils/storybook';
 import { ACCOUNT_STATUS_PAGE } from '../account/index.stories';
 import RejectRequest from './components/RejectRequest';
 import ShowRequest from './components/ShowRequest';
+import { getTransaction } from './test';
 
 const TX_REQUEST_PATH = `${CHROME_EXTENSION_ROOT}/Transaction Request`;
 const SHOW_REQUEST_PAGE = 'Show Request page';
 const REJECT_REQUEST_PAGE = 'Reject Request page';
-
-const defaultCreator: PublicIdentity = {
-  chainId: 'some-chain' as ChainId,
-  pubkey: {
-    algo: Algorithm.Ed25519,
-    // Random 32 bytes pubkey. Derived IOV address:
-    // tiov1dcg3fat5zrvw00xezzjk3jgedm7pg70y222af3 / 6e1114f57410d8e7bcd910a568c9196efc1479e4
-    data: Encoding.fromHex(
-      '7196c465e4c95b3dce425784f51936b95da6bc58b3212648cdca64ee7198df47',
-    ) as PublicKeyBytes,
-  },
-};
-
-const defaultAmount: Amount = {
-  quantity: '1000000001',
-  fractionalDigits: 9,
-  tokenTicker: 'CASH' as TokenTicker,
-};
-
-const transaction: SendTransaction = {
-  kind: 'bcp/send',
-  creator: defaultCreator,
-  amount: defaultAmount,
-  recipient: 'tiov1k898u78hgs36uqw68dg7va5nfkgstu5z0fhz3f' as Address,
-  memo: 'paid transaction',
-  fee: {
-    tokens: defaultAmount,
-  },
-};
 
 const request: Request = {
   id: 0,
@@ -63,7 +24,7 @@ const request: Request = {
   reason: 'I would like you to sign this TX',
   data: {
     senderUrl: 'http://localhost/',
-    tx: transaction,
+    tx: getTransaction(),
   },
   type: 'signAndPost',
 };
